refactor: Re-export entities
