+++ backend_rust/src/adapters/db/db_recipe_repository.rs
@@ -6,8 +6,8 @@ use crate::adapters::db::mysql::entity as db_entity;
 use crate::adapters::recipe_mapper::RecipeMapper;
 use crate::adapters::step_mapper::StepMapper;
 use crate::application::mapper::db_mapper::DbMapper;
-use crate::domain::recipe::Recipe;
-use crate::domain::step::Step;
+use crate::domain::Recipe;
+use crate::domain::Step;
 use async_trait::async_trait;
 use sea_orm::*;
