+++ backend_rust/src/adapters/db/db_resource_repository.rs
@@ -1,5 +1,6 @@
 use crate::adapters::db::mysql::entity as db_entity;
 use crate::application::repository::resource_repository::ResourceRepository;
+use crate::domain::Resource;
 use async_trait::async_trait;
 use sea_orm::DatabaseConnection;
 use sea_orm::*;
@@ -11,17 +12,14 @@ pub struct DbResourceRepository {
 
 #[async_trait]
 impl ResourceRepository for DbResourceRepository {
-    async fn get_resource_by_id(
-        &self,
-        id: i32,
-    ) -> Result<crate::domain::resource::Resource, Box<dyn std::error::Error>> {
+    async fn get_resource_by_id(&self, id: i32) -> Result<Resource, Box<dyn std::error::Error>> {
         let model = db_entity::resources::Entity::find_by_id(id as u64)
             .one(&*self.db_connection)
             .await;
 
         match model {
             Ok(Some(model)) => {
-                let resource = crate::domain::resource::Resource {
+                let resource = Resource {
                     id: model.id as i32,
                     name: model.name,
                     amount: model.amount,
@@ -39,9 +37,7 @@ impl ResourceRepository for DbResourceRepository {
         }
     }
 
-    async fn get_all_resources(
-        &self,
-    ) -> Result<Vec<crate::domain::resource::Resource>, Box<dyn std::error::Error>> {
+    async fn get_all_resources(&self) -> Result<Vec<Resource>, Box<dyn std::error::Error>> {
         let models = db_entity::resources::Entity::find()
             .all(&*self.db_connection)
             .await;
@@ -50,12 +46,12 @@ impl ResourceRepository for DbResourceRepository {
             Ok(models) => {
                 let models = models
                     .into_iter()
-                    .map(|model| crate::domain::resource::Resource {
+                    .map(|model| Resource {
                         id: model.id as i32,
                         name: model.name,
                         amount: model.amount,
                     })
-                    .collect::<Vec<crate::domain::resource::Resource>>();
+                    .collect::<Vec<Resource>>();
                 Ok(models)
             }
             Err(e) => Err(Box::new(std::io::Error::other(e))),
