+++ backend_rust/src/adapters/step_mapper.rs
@@ -1,8 +1,8 @@
 use crate::adapters::db::mysql::entity::steps::Model as StepModel;
 use crate::application::mapper::api_mapper::ApiMapper;
 use crate::application::mapper::db_mapper::DbMapper;
-use crate::domain::resource::Resource;
-use crate::domain::step::Step;
+use crate::domain::Resource;
+use crate::domain::Step;
 use crate::presentation::graphql::object::Step as StepObject;
 pub struct StepMapper {}
