+++ backend_rust/src/domain/entity.rs
@@ -1,0 +1,4 @@
+mod process;
+pub mod recipe;
+pub mod resource;
+pub mod step;
