+++ backend_rust/src/domain/entity/resource.rs
